feat(tests): refactor utils into manifest_utils mod
