+++ tests/manifest_utils/mod.rs
